fileshare

+++ database/migrations/2020_05_06_010743_create_files_table.php
@@ -16,23 +16,16 @@ class CreateFilesTable extends Migration
     {
         Schema::create('files', function (Blueprint $table) {
             $table->id();
-
             $table->bigInteger('user_id')->unsigned();
-
             $table->string('name');
-
             $table->string('hash_name');
-
+            $table->string('path');
             $table->bigInteger('size')->unsigned();
-
-            $table->text('comment')->nullable();
-
             $table->string('extension');
-
             $table->timestamps();
             $table->softDeletes();
 
-            $table->foreign('user_id')->references('id')->on('users');
+            $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
         });
     }
 
@@ -44,13 +37,17 @@ class CreateFilesTable extends Migration
     public function down()
     {
         $items = File::all();
-        foreach($items as $item)
+        $files = [];
+        foreach ($items as $item)
         {
             //unlink("public/storage/uploads/user_2/" . $item->hash_name);
-            $files[] = "uploads/user_2/" . $item->hash_name;
+            $path = "uploads/user_" . $item->user_id . "/" . $item->hash_name;
+            echo $path . "\n";
+            $files[] = $path;
         }
         //пизда ебаная не работала я думал ебнусь
-        \Storage::disk('public')->delete($files);
+        if($files)
+            \Storage::disk('public')->delete($files);
 
         Schema::dropIfExists('files');
     }
